requirements.txt
flask
pandas
torch
torchvision
open-clip-torch
Pillow
scikit-learn
